hw-9 delete_id

+++ src/routes/security-router.ts
@@ -80,6 +80,8 @@ export const securityRouter = () => {
       const currentDeviceId = req.deviceId
       const userId = req.user?.id
       if (userId && currentDeviceId) {
+        const user = await devicesRepository.findById(userId)
+        if (!user) return res.sendStatus(404)
         const deletedDevices = await devicesService.deleteDevicesWithoutCurrent(
           userId,
           currentDeviceId
